+++ src/scripts/updateSimfileSheet.js
@@ -1,5 +1,7 @@
 const fs = require("fs");
+
 const simfileTsvPath = "../../public/data/simfiles.tsv";
+const { getMetadataFromSM } = require("./smParser");
 
 const parseSimfileTsv = () => {
   let tsv = fs.readFileSync(simfileTsvPath, "utf-8").split("\n");
@@ -10,6 +12,7 @@ const parseSimfileTsv = () => {
   for (let i = 1; i < tsv.length; i++) {
     const values = tsv[i].split("\t");
     const row = {};
+    row.index = i - 1; // way to keep track of the original eamuse song order
     values.forEach((value, col) => {
       if (value === "") value = null;
       const header = headers[col];
@@ -21,7 +24,7 @@ const parseSimfileTsv = () => {
   return output;
 };
 
-const writeSimfileTsv = json => {
+const writeSimfileTsv = (json) => {
   let output = "";
 
   console.log(json);
@@ -29,9 +32,9 @@ const writeSimfileTsv = json => {
 
   output += headers.join("\t");
 
-  json.forEach(song => {
+  json.forEach((song) => {
     let row = "\n";
-    headers.forEach(header => {
+    headers.forEach((header) => {
       let value = song[header];
       if (value === null) value = "";
       row += value + "\t";
@@ -41,47 +44,48 @@ const writeSimfileTsv = json => {
 
   console.log(output);
 
-  fs.writeFile(simfileTsvPath, output, "utf8", err => {
+  fs.writeFile(simfileTsvPath, output, "utf8", (err) => {
     if (err) console.log(err);
   });
 };
 
-// one-time script
-const getMp3AndSmColumns = () => {
-  return fs
-    .readFileSync("dropbox_mp3_urls.txt", "utf-8")
-    .split("\n")
-    .map(url => {
-      const dAudioUrl = url
-        .replace(`https://www.dropbox.com/s/`, "")
-        .replace("?dl=0", "");
-      const smName = decodeURI(
-        dAudioUrl.slice(
-          dAudioUrl.indexOf("/") + 1,
-          dAudioUrl.indexOf("mp3") - 1
-        )
-      );
-
-      return { dAudioUrl, smName };
-    });
-};
-
 const init = async () => {
-  const parsedTsv = parseSimfileTsv();
+  let parsedTsv = parseSimfileTsv();
 
-  // one-time script of filling in empty smName and mp3 url columns
-  const newColumns = getMp3AndSmColumns();
-  parsedTsv.forEach(song => {
-    if (!song.smName) return;
-    song.smName = song.smName.replace(".sm", "");
-    for (let column of newColumns) {
-      if (song.smName === column.smName) {
-        song.dAudioUrl = column.dAudioUrl;
-      }
-    }
-  });
+  parsedTsv = getMetadataFromSM(parsedTsv);
+
+  // // one-time script of filling in empty smName and mp3 url columns
+  // const newColumns = getMp3AndSmColumns();
+  // parsedTsv.forEach(song => {
+  //   if (!song.smName) return;
+  //   song.smName = song.smName.replace(".sm", "");
+  //   for (let column of newColumns) {
+  //     if (song.smName === column.smName) {
+  //       song.dAudioUrl = column.dAudioUrl;
+  //     }
+  //   }
+  // });
 
   await writeSimfileTsv(parsedTsv);
 };
 
 init();
+
+// const getMp3AndSmColumns = () => {
+//   return fs
+//     .readFileSync("dropbox_mp3_urls.txt", "utf-8")
+//     .split("\n")
+//     .map(url => {
+//       const dAudioUrl = url
+//         .replace(`https://www.dropbox.com/s/`, "")
+//         .replace("?dl=0", "");
+//       const smName = decodeURI(
+//         dAudioUrl.slice(
+//           dAudioUrl.indexOf("/") + 1,
+//           dAudioUrl.indexOf("mp3") - 1
+//         )
+//       );
+
+//       return { dAudioUrl, smName };
+//     });
+// };